fat-use-fat_fs_error-instead-of-bug_on-in-__fat_get_block-fix
fix printk warning

Link: http://lkml.kernel.org/r/87po12aq5p.fsf@mail.parknet.co.jp
Cc: Anatoly Trosinenko <anatoly.trosinenko@gmail.com>
Tested-by: Anatoly Trosinenko <anatoly.trosinenko@gmail.com>
Cc: Arnd Bergmann <arnd@arndb.de>

+++ fs/fat/inode.c
@@ -160,8 +160,9 @@ static inline int __fat_get_block(struct inode *inode, sector_t iblock,
 		return err;
 	if (!phys) {
 		fat_fs_error(sb,
-			     "invalid FAT chain (i_pos %lld, last_block %ld)",
-			     MSDOS_I(inode)->i_pos, last_block);
+			     "invalid FAT chain (i_pos %lld, last_block %llu)",
+			     MSDOS_I(inode)->i_pos,
+			     (unsigned long long)last_block);
 		return -EIO;
 	}
